refactor: order, review, store
