+++ src/main/java/com/sparta/outsourcing/domain/order/dto/LogResponseDto.java
@@ -1,0 +1,18 @@
+package com.sparta.outsourcing.domain.order.dto;
+
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class LogResponseDto {
+    private String timestamp;
+    private Long storeId;
+    private Long orderId;
+
+    public LogResponseDto(Long orderId, Long storeId) {
+        this.timestamp = LocalDateTime.now().toString();
+        this.storeId = storeId;
+        this.orderId = orderId;
+    }
+}
