+++ src/main/java/com/sparta/outsourcing/domain/order/dto/OrderRequestDto.java
@@ -1,0 +1,9 @@
+package com.sparta.outsourcing.domain.order.dto;
+
+import lombok.Getter;
+
+@Getter
+public class OrderRequestDto {
+    private String menu;
+    private int totalPrice;
+}
