+++ src/main/java/com/sparta/outsourcing/domain/order/enums/OrderStatusEnum.java
@@ -1,0 +1,8 @@
+package com.sparta.outsourcing.domain.order.enums;
+
+public enum OrderStatusEnum {
+    NEW,
+    ACCEPTED,
+    DELIVERED,
+    CANCELLED
+}
