+++ src/main/java/com/sparta/outsourcing/domain/order/orderstatusenum/OrderStatus.java
@@ -1,0 +1,8 @@
+package com.sparta.outsourcing.domain.order.orderstatusenum;
+
+public enum OrderStatus {
+    NEW,
+    ACCEPTED,
+    DELIVERED,
+    CANCELLED
+}
