+++ src/main/java/com/sparta/outsourcing/domain/review/reviewDTO/ReviewRequestDto.java
@@ -1,0 +1,15 @@
+package com.sparta.outsourcing.domain.review.reviewDTO;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ReviewRequestDto {
+
+    private Long reviewId;
+    private Long orderId;
+    private int rating;
+    private String contents;
+    private String reviewPictureUrl;
+}
