+++ src/main/java/com/sparta/outsourcing/domain/stores/enums/StoreStatus.java
@@ -1,0 +1,6 @@
+package com.sparta.outsourcing.domain.stores.enums;
+
+public enum StoreStatus {
+    Deleted,
+    Running
+}
