--- src/main/java/com/sparta/outsourcing/order/dto/OrderRequestDto.java
@@ -1,9 +1,0 @@
-package com.sparta.outsourcing.order.dto;
-
-import lombok.Getter;
-
-@Getter
-public class OrderRequestDto {
-    private String menu;
-    private int totalPrice;
-}
