--- src/main/java/com/sparta/outsourcing/order/enums/OrderStatusEnum.java
@@ -1,8 +1,0 @@
-package com.sparta.outsourcing.order.enums;
-
-public enum OrderStatusEnum {
-    NEW,
-    ACCEPTED,
-    DELIVERED,
-    CANCELLED
-}
