--- src/main/java/com/sparta/outsourcing/order/orderstatusenum/OrderStatus.java
@@ -1,8 +1,0 @@
-package com.sparta.outsourcing.order.orderstatusenum;
-
-public enum OrderStatus {
-    NEW,
-    ACCEPTED,
-    DELIVERED,
-    CANCELLED
-}
