--- src/main/java/com/sparta/outsourcing/order/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.sparta.outsourcing.order.repository;
-
-import com.sparta.outsourcing.order.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
