--- src/main/java/com/sparta/outsourcing/review/reviewDTO/ReviewRequestDto.java
@@ -1,15 +1,0 @@
-package com.sparta.outsourcing.review.reviewDTO;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ReviewRequestDto {
-
-    private Long reviewId;
-    private Long orderId;
-    private int rating;
-    private String contents;
-    private String reviewPictureUrl;
-}
