--- src/main/java/com/sparta/outsourcing/stores/dto/StoreResponseDto.java
@@ -1,16 +1,0 @@
-package com.sparta.outsourcing.stores.dto;
-
-import lombok.Getter;
-
-@Getter
-public class StoreResponseDto {
-    private String message;
-    private String name;
-    private Integer status;
-
-    public StoreResponseDto(String message, String name, Integer status) {
-        this.message = message;
-        this.name = name;
-        this.status = status;
-    }
-}
