--- src/main/java/com/sparta/outsourcing/stores/enums/StoreStatus.java
@@ -1,6 +1,0 @@
-package com.sparta.outsourcing.stores.enums;
-
-public enum StoreStatus {
-    Deleted,
-    Running
-}
