--- src/main/java/com/sparta/outsourcing/stores/repository/StoresRepository.java
@@ -1,7 +1,0 @@
-package com.sparta.outsourcing.stores.repository;
-
-import com.sparta.outsourcing.stores.entity.Stores;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoresRepository extends JpaRepository<Stores, Long> {
-}
